# packages/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bessel API",
    "summary": "Bessel HTTP and Webhooks API",
    "description": "Hello! This is the Bessel API.",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "https://api.getbessel.com",
      "description": "Production environment",
      "x-speakeasy-server-id": "production"
    }
  ],
  "paths": {
...
    "/v1/readings": {
      "post": {
        "tags": [
          "readings"
        ],
        "summary": "Capture Reading",
        "operationId": "create_reading_v1_readings_post",
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadingSchema"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "readings"
        ],
        "summary": "List Readings",
        "operationId": "list_readings_v1_readings_get",
        "parameters": [
          {
            "name": "created_after",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter readings created after this Unix timestamp (inclusive).",
              "title": "Created After"
            },
            "description": "Filter readings created after this Unix timestamp (inclusive)."
          },
          {
            "name": "created_before",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter readings created before this Unix timestamp (exclusive).",
              "title": "Created Before"
            },
            "description": "Filter readings created before this Unix timestamp (exclusive)."
          },
          {
            "name": "has_alarms",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by alarm presence.",
              "title": "Has Alarms"
            },
            "description": "Filter by alarm presence."
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "description": "Page number, defaults to 1.",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number, defaults to 1."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "description": "Size of a page, defaults to 10. Maximum is 100.",
              "default": 10,
              "title": "Limit"
            },
            "description": "Size of a page, defaults to 10. Maximum is 100."
          },
          {
            "name": "sorting",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReadingSortProperty"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Sorting criterion. Several criteria can be used simultaneously and will be applied in order. Add a minus sign `-` before the criteria name to sort by descending order.",
              "default": [
                "-created_at"
              ],
              "title": "Sorting"
            },
            "description": "Sorting criterion. Several criteria can be used simultaneously and will be applied in order. Add a minus sign `-` before the criteria name to sort by descending order."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadingListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/readings/{reading_id}": {
      "get": {
        "tags": [
          "readings"
        ],
        "summary": "Get Reading",
        "operationId": "get_reading_v1_readings__reading_id__get",
        "parameters": [
          {
            "name": "reading_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Reading Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadingSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BalancingSchema": {
        "properties": {
          "balance_current_a": {
            "type": "number",
            "title": "Balance Current A",
            "description": "Balancer current in amperes."
          },
          "balancing_cells": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Balancing Cells",
            "description": "1-based numbers of the cells currently being balanced."
          },
          "mosfet_temperature_c": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mosfet Temperature C",
            "description": "Balancer MOSFET temperature in \u00b0C."
          },
          "board_temperature_c": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Board Temperature C",
            "description": "Balancer board temperature in \u00b0C."
          }
        },
        "type": "object",
        "required": [
          "balance_current_a",
          "balancing_cells"
        ],
        "title": "BalancingSchema"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Pagination": {
        "properties": {
          "total_count": {
            "type": "integer",
            "title": "Total Count"
          },
          "max_page": {
            "type": "integer",
            "title": "Max Page"
          }
        },
        "type": "object",
        "required": [
          "total_count",
          "max_page"
        ],
        "title": "Pagination"
      },
      "ReadingListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ReadingSchema"
            },
            "type": "array",
            "title": "Items"
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        },
        "type": "object",
        "required": [
          "items",
          "pagination"
        ],
        "title": "ReadingListResponse"
      },
      "ReadingSchema": {
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Creation timestamp of the object."
          },
          "modified_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Modified At",
            "description": "Last modification timestamp of the object."
          },
          "id": {
            "type": "string",
            "format": "uuid4",
            "title": "Id",
            "description": "The ID of the object."
          },
          "device_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Name",
            "description": "Bluetooth name of the BMS device."
          },
          "device_address": {
            "type": "string",
            "title": "Device Address",
            "description": "Bluetooth address of the BMS device."
          },
          "device_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Model",
            "description": "BMS model string reported by the device."
          },
          "protocol": {
            "type": "string",
            "title": "Protocol",
            "description": "Protocol used to read the BMS."
          },
          "voltage_v": {
            "type": "number",
            "title": "Voltage V",
            "description": "Total pack voltage in volts."
          },
          "current_a": {
            "type": "number",
            "title": "Current A",
            "description": "Pack current in amperes; positive when charging, negative when discharging."
          },
          "power_w": {
            "type": "number",
            "title": "Power W",
            "description": "Pack power in watts."
          },
          "soc_percent": {
            "type": "number",
            "title": "Soc Percent",
            "description": "State of charge in percent."
          },
          "remaining_capacity_ah": {
            "type": "number",
            "title": "Remaining Capacity Ah",
            "description": "Remaining capacity in ampere-hours."
          },
          "rated_capacity_ah": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rated Capacity Ah",
            "description": "Rated capacity in ampere-hours, from the BMS settings."
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "description": "Pack mode: idle, charging or discharging."
          },
          "cycles": {
            "type": "integer",
            "title": "Cycles",
            "description": "Charge cycle count."
          },
          "cell_count": {
            "type": "integer",
            "title": "Cell Count",
            "description": "Number of cells in the pack."
          },
          "cell_voltage_max_v": {
            "type": "number",
            "title": "Cell Voltage Max V",
            "description": "Highest cell voltage in volts."
          },
          "cell_voltage_min_v": {
            "type": "number",
            "title": "Cell Voltage Min V",
            "description": "Lowest cell voltage in volts."
          },
          "cell_voltage_avg_v": {
            "type": "number",
            "title": "Cell Voltage Avg V",
            "description": "Average cell voltage in volts."
          },
          "cell_voltage_delta_v": {
            "type": "number",
            "title": "Cell Voltage Delta V",
            "description": "Difference between highest and lowest cell voltage in volts."
          },
          "temperature_sensor_count": {
            "type": "integer",
            "title": "Temperature Sensor Count",
            "description": "Number of temperature sensors."
          },
          "temperature_max_c": {
            "type": "number",
            "title": "Temperature Max C",
            "description": "Highest sensor temperature in \u00b0C."
          },
          "temperature_min_c": {
            "type": "number",
            "title": "Temperature Min C",
            "description": "Lowest sensor temperature in \u00b0C."
          },
          "balancer_active": {
            "type": "boolean",
            "title": "Balancer Active",
            "description": "Whether the balancer is active."
          },
          "charging_mosfet": {
            "type": "boolean",
            "title": "Charging Mosfet",
            "description": "Whether the charging MOSFET is enabled."
          },
          "discharging_mosfet": {
            "type": "boolean",
            "title": "Discharging Mosfet",
            "description": "Whether the discharging MOSFET is enabled."
          },
          "has_alarms": {
            "type": "boolean",
            "title": "Has Alarms",
            "description": "Whether any alarm register is set."
          },
          "cell_voltages": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Cell Voltages",
            "description": "Per-cell voltages in volts, ordered by cell number."
          },
          "temperatures": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Temperatures",
            "description": "Per-sensor temperatures in \u00b0C, ordered by sensor number."
          },
          "balancing": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BalancingSchema"
              },
              {
                "type": "null"
              }
            ],
            "description": "Balancing details, if reported by the BMS."
          },
          "alarms": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alarms",
            "description": "Raw alarm registers as hex strings, or null when no alarms are set."
          }
        },
        "type": "object",
        "required": [
          "created_at",
          "modified_at",
          "id",
          "device_name",
          "device_address",
          "device_model",
          "protocol",
          "voltage_v",
          "current_a",
          "power_w",
          "soc_percent",
          "remaining_capacity_ah",
          "rated_capacity_ah",
          "mode",
          "cycles",
          "cell_count",
          "cell_voltage_max_v",
          "cell_voltage_min_v",
          "cell_voltage_avg_v",
          "cell_voltage_delta_v",
          "temperature_sensor_count",
          "temperature_max_c",
          "temperature_min_c",
          "balancer_active",
          "charging_mosfet",
          "discharging_mosfet",
          "has_alarms",
          "cell_voltages",
          "temperatures",
          "balancing",
          "alarms"
        ],
        "title": "ReadingSchema"
      },
      "ReadingSortProperty": {
        "type": "string",
        "enum": [
          "created_at",
          "-created_at",
          "voltage_v",
          "-voltage_v",
          "soc_percent",
          "-soc_percent"
        ],
        "title": "ReadingSortProperty"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}